refactor: moved app.py helper functions to src package
