+++ alarmbot/src/__init__.py
@@ -1,1 +1,1 @@
-
+from process_request import process_request
