+++ alarmbot/src/process_request/__init__.py
@@ -1,0 +1,1 @@
+from process_request import run
