+++ alarmbot/src/val_api/__init__.py
@@ -1,0 +1,1 @@
+from val_api import help, default, get_rank, get_recent_summary, get_recent_full
